requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
sqlalchemy==2.0.40
psycopg[binary]==3.2.13
redis==5.2.1
httpx==0.28.1
pydantic-settings==2.9.1
python-dotenv==1.1.0
pytest==8.3.5
pytest-asyncio==0.26.0
